Add Frameworks drawable-hdpi images into the overlay.
according to device.xml system will still fall back on hdpi if the xhdpi image doesn't exist. Because of this have added all “drawable-hdpi” images ware from the last ASOP based incarnation of the Starship theme to get an idea of how frequent if actually at all this will happen and if it will be necessary to cover both versions.